+++ SignNow.Net.Test/AcceptanceTests/DocumentServiceTest.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Collections.Generic;
+using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SignNow.Net.Internal.Extensions;
 using SignNow.Net.Model;
@@ -43,6 +45,27 @@ namespace AcceptanceTests
 
             Assert.AreEqual("merged-document.pdf", merged.Filename);
             Assert.That.StreamIsPdf(merged.Document);
+        }
+
+        [TestMethod]
+        public void DocumentHistory()
+        {
+            var response = SignNowTestContext.Documents
+                .GetDocumentHistoryAsync(TestPdfDocumentIdWithFields)
+                .Result;
+
+            Assert.IsTrue(response.All(itm => itm.Id.Length == 40));
+            Assert.IsTrue(response.All(itm => itm.DocumentId == TestPdfDocumentIdWithFields));
+        }
+
+        [TestMethod]
+        public void CreateOneTimeDocumentDownloadLink()
+        {
+            var link = SignNowTestContext.Documents
+                .CreateOneTimeDownloadLinkAsync(TestPdfDocumentId).Result;
+
+            Assert.IsNotNull(link.Url);
+            StringAssert.Contains(link.Url.Host, "signnow.com");
         }
     }
 }