+++ SignNow.Net/Interfaces/IDocumentService.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using SignNow.Net.Model;
 using System.IO;
@@ -7,7 +8,8 @@ using System.Threading.Tasks;
 namespace SignNow.Net.Interfaces
 {
     /// <summary>
-    /// Interface for any operations with a Document in SignNow; can be used to create, download, retrieve, delete a document etc.
+    /// Interface for any operations with a Document in SignNow
+    /// can be used to create, download, retrieve, delete a document etc.
     /// </summary>
     public interface IDocumentService
     {
@@ -21,7 +23,8 @@ namespace SignNow.Net.Interfaces
         Task<SignNowDocument> GetDocumentAsync(string documentId, CancellationToken cancellationToken = default);
 
         /// <summary>
-        /// Uploads a file to the SignNow account and creates a document. This method accepts .doc, .docx, .pdf, and .png file types.
+        /// Uploads a file to the SignNow account and creates a document.
+        /// This method accepts .doc, .docx, .pdf, and .png file types.
         /// </summary>
         /// <param name="documentContent">Document content stream</param>
         /// <param name="fileName">Uploaded document file name</param>
@@ -32,13 +35,12 @@ namespace SignNow.Net.Interfaces
         /// <summary>
         /// Uploads a file to the SignNow account, creates a document and extracts simple field tags if any.
         /// This method accepts .doc, .docx and .pdf file types.
-        /// See <a href="https://campus.barracuda.com/product/signnow/doc/41113461/rest-endpoints-api">REST API Endpoints</a>
         /// </summary>
         /// <param name="documentContent">Document content stream</param>
         /// <param name="fileName">Uploaded document file name</param>
         /// <param name="cancellationToken">Propagates notification that operations should be canceled.</param>
         /// <returns>Operation result object containing ID of the new document.</returns>
-        Task<UploadDocumentResponse> UploadDocumentWithFieldExtractAsync(Stream documentContent, string fileName, CancellationToken cancellationToken = default);
+        Task<UploadDocumentResponse>UploadDocumentWithFieldExtractAsync(Stream documentContent, string fileName, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Removes a document from SignNow account.
@@ -77,5 +79,24 @@ namespace SignNow.Net.Interfaces
         /// <returns>Operation result object containing File info with Stream file content.</returns>
         /// <exception cref="System.ArgumentException">If document identity is not valid.</exception>
         Task<DownloadDocumentResponse> MergeDocumentsAsync(string documentName, IEnumerable<SignNowDocument> documents, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Retrieves a document history (action log) data
+        /// </summary>
+        /// <param name="documentId">Identity of the document to retrieve history list</param>
+        /// <param name="cancellationToken">Propagates notification that operations should be canceled</param>
+        /// <returns>List of Document history items</returns>
+        /// <exception cref="System.ArgumentException">If document identity is not valid.</exception>
+        Task<IReadOnlyList<DocumentHistoryResponse>> GetDocumentHistoryAsync(string documentId, CancellationToken cancellationToken = default);
+
+
+        /// <summary>
+        /// Create a one-time use URL for anyone to download the document as a PDF.
+        /// </summary>
+        /// <param name="documentId">Identity of the document to be downloaded with one-time link</param>
+        /// <param name="cancellationToken">Propagates notification that operations should be canceled</param>
+        /// <returns>link to download specified document in PDF format</returns>
+        /// <exception cref="System.ArgumentException">If document identity is not valid.</exception>
+        Task<DownloadLinkResponse> CreateOneTimeDownloadLinkAsync(string documentId, CancellationToken cancellationToken = default);
     }
 }